refactor(domain): move validation files to subfolder
