+++ src/core/domain/validation/ValidationWarningType.ts
