+++ src/ui/hooks/useValidationWarnings.ts
@@ -2,7 +2,7 @@ import { useMemo } from 'react';
 import { useBoard, useLinks } from '../../core/store/useBoardStore';
 import { ValidateModelQuery } from '../../core/usecases/queries/ValidateModel/ValidateModelQuery';
 import { ValidateModelQueryHandler } from '../../core/usecases/queries/ValidateModel/ValidateModelQueryHandler';
-import { type ValidationWarning } from '../../core/domain/ValidationWarning';
+import { type ValidationWarning } from '../../core/domain/validation/ValidationWarning';
 
 export function useValidationWarnings(): ReadonlyArray<ValidationWarning> {
   const board = useBoard();
